Stun baton prediction
