+++ Content.Shared/Stunnable/StunbatonComponent.cs
@@ -1,8 +1,7 @@
-using Content.Shared.Stunnable;
 using Robust.Shared.Audio;
 using Robust.Shared.GameStates;
 
-namespace Content.Server.Stunnable.Components;
+namespace Content.Shared.Stunnable.Components;
 
 [RegisterComponent, NetworkedComponent]
 [AutoGenerateComponentState]
